# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Swish Stats — basketball stat capture API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues:
    get:
      operationId: listLeagues
      tags: [leagues]
      summary: List all leagues the current user can access
      responses:
        "200":
          description: List of leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
    post:
      operationId: createLeague
      tags: [leagues]
      summary: Create a new league (current user becomes admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeagueBody"
      responses:
        "201":
          description: League created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /leagues/{leagueId}:
    get:
      operationId: getLeague
      tags: [leagues]
      summary: Get a single league
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateLeague
      tags: [leagues]
      summary: Update a league (admin only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeagueBody"
      responses:
        "200":
          description: Updated league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteLeague
      tags: [leagues]
      summary: Delete a league (admin only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /leagues/{leagueId}/games:
    get:
      operationId: listLeagueGames
      tags: [leagues]
      summary: List games belonging to a league
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createLeagueGame
      tags: [leagues]
      summary: Create a new game inside a league (admin only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameBody"
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /leagues/{leagueId}/members:
    get:
      operationId: listLeagueMembers
      tags: [memberships]
      summary: List memberships for a league
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeagueMember"
    post:
      operationId: addLeagueMember
      tags: [memberships]
      summary: Add a member to a league (admin only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLeagueMemberBody"
      responses:
        "201":
          description: Added membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueMember"
  /leagues/{leagueId}/members/{userId}:
    patch:
      operationId: updateLeagueMember
      tags: [memberships]
      summary: Change a member's role (admin only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeagueMemberBody"
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueMember"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeLeagueMember
      tags: [memberships]
      summary: Remove a member from a league (admin only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: integer
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    LeagueRole:
      type: string
      enum: [viewer, scorer, admin]
    League:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        season:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        ownerUserId:
          type: ["string", "null"]
        viewerRole:
          $ref: "#/components/schemas/LeagueRole"
        createdAt:
          type: string
          format: date-time
      required: [id, name, viewerRole, createdAt]
    CreateLeagueBody:
      type: object
      properties:
        name:
          type: string
        season:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
      required: [name]
    UpdateLeagueBody:
      type: object
      properties:
        name:
          type: string
        season:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
    LeagueMember:
      type: object
      properties:
        leagueId:
          type: integer
        userId:
          type: string
        role:
          $ref: "#/components/schemas/LeagueRole"
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [leagueId, userId, role, createdAt]
    AddLeagueMemberBody:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/LeagueRole"
      required: [email, role]
    UpdateLeagueMemberBody:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/LeagueRole"
      required: [role]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    GameStatus:
      type: string
      enum: [setup, active, final]
    CaptureMode:
      type: string
      enum: [simple, complex]
    Game:
      type: object
      properties:
        id:
          type: integer
        leagueId:
          type: ["integer", "null"]
        competition:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        venue:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/GameStatus"
        captureMode:
          $ref: "#/components/schemas/CaptureMode"
        periodCount:
          type: integer
        periodDurationMins:
          type: integer
        overtimeCount:
          type: integer
        currentPeriod:
          type: integer
        clockSeconds:
          type: integer
        possessionTeamId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - status
        - captureMode
        - periodCount
        - periodDurationMins
        - overtimeCount
        - currentPeriod
        - clockSeconds
        - createdAt
        - updatedAt
    CreateGameBody:
      type: object
      properties:
        leagueId:
          type: ["integer", "null"]
        competition:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        venue:
          type: ["string", "null"]
        captureMode:
          $ref: "#/components/schemas/CaptureMode"
        periodCount:
          type: integer
        periodDurationMins:
          type: integer
...