# lib/api-zod/src/generated/types/updateLeagueMemberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swish Stats — basketball stat capture API
 * OpenAPI spec version: 0.1.0
 */
import type { LeagueRole } from "./leagueRole";

export interface UpdateLeagueMemberBody {
  role: LeagueRole;
}